+++ tools/config.vala
@@ -26,7 +26,19 @@ const string ENGINE_EXE = EXE_PREFIX
 const string DEPLOY_DEFAULT_NAME = "crown-release";
 const string DEPLOY_EXE = EXE_PREFIX + DEPLOY_DEFAULT_NAME + EXE_SUFFIX;
 
-const uint16 CROWN_DEFAULT_SERVER_PORT = 10618;
+const uint16 DATA_COMPILER_TCP_PORT = 10618;
+const uint16 EDITOR_TCP_PORT = 10001;
+const uint16 UNIT_PREVIEW_TCP_PORT = 10002;
+const uint16 GAME_TCP_PORT = 12345;
+
+const int DATA_COMPILER_CONNECTION_TRIES = 10;
+const int DATA_COMPILER_CONNECTION_INTERVAL = 250;
+
+const int EDITOR_CONNECTION_TRIES = 16;
+const int EDITOR_CONNECTION_INTERVAL = 125;
+
+const int GAME_CONNECTION_TRIES = 16;
+const int GAME_CONNECTION_INTERVAL = 125;
 
 const string LEVEL_EDITOR_BOOT_DIR = "core/editors/level_editor";
 const string UNIT_PREVIEW_BOOT_DIR = "core/editors/unit_preview";